Deprecate sync connection manager
Move connection manager code into `deprecated` module, so that the
`connection_manager` module can contain new async code.
