+++ comms/src/connection_manager/deprecated/actor.rs
@@ -1,4 +1,4 @@
-// Copyright 2019 The Tari Project
+// Copyright 2019, The Tari Project
 //
 // Redistribution and use in source and binary forms, with or without modification, are permitted provided that the
 // following conditions are met:
@@ -20,11 +20,8 @@
 // WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE
 // USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-use crate::{
-    connection::PeerConnection,
-    connection_manager::{dialer::Dialer, ConnectionManagerError, Connectivity},
-    peer_manager::NodeId,
-};
+use super::{dialer::Dialer, ConnectionManagerError, Connectivity};
+use crate::{connection::PeerConnection, peer_manager::NodeId};
 use futures::{
     channel::{mpsc, oneshot},
     future::{self, BoxFuture, Either, FutureExt},
