+++ comms/src/connection_manager/deprecated/connectivity.rs
